requirements.txt
discord.py
psutil
requests
gitpython
pymongo
dnspython
circleci
python-dotenv
fake-useragent
beautifulsoup4